Refactor FOV, scripting, audio & assorted fixes
Large cleanup and functional changes across the codebase:

- FOV/aspect handling: docs updated; deprecated r_aspectRatio removed; per-frame/gameplay FOV now derived from actual render size. Removed legacy aspect-agnostic zoom helpers and simplified weapon/player zoom paths. idGameLocal::CalcFov rewritten to use actual render resolution.

- Scripting & threading: hhThread API updated to use intptr_t for PushParm and improved PushString/PushVector/PushInt behavior; ParseAndPushArgsOntoStack tightened. Header updated accordingly.

- Audio: spawn sound parsing now uses raw s_volume (no dB->scale conversion); StartFade implementation simplified/stubbed and callers adjusted to guard volume access. Other sound-related cleanup.

- Gameplay & entity fixes: numerous const-correctness, null checks and small bug fixes (e.g. BroadcastFxInfoAlongBonePrefix, SetTargetEntity call, subtitle/gui handling, mover/pusher behavior).

- Misc: added friend declaration in AASFile, updated GAME_ICON path, compacted some enums and formatting/whitespace cleanups, and replaced legacy OpenPrey entity/spawn helpers with a safer OpenPrey_GetSpawnIdForEntity implementation.

These changes focus on correctness, safety, and aligning runtime behavior with current renderer/state instead of legacy assumptions.

+++ src/game/gamesys/Class.h
@@ -10,6 +10,8 @@ instancing of objects.
 #ifndef __SYS_CLASS_H__
 #define __SYS_CLASS_H__
 
+#include <stdint.h>
+
 class idClass;
 class idTypeInfo;
 
@@ -35,8 +37,8 @@ public:
 	intptr_t	value;
 
 	idEventArg()								{ type = D_EVENT_INTEGER; value = 0; };
-	idEventArg( int data )						{ type = D_EVENT_INTEGER; value = data; };
-	idEventArg( float data )					{ type = D_EVENT_FLOAT; value = *reinterpret_cast<int *>( &data ); };
+	idEventArg( int data )						{ type = D_EVENT_INTEGER; value = static_cast<intptr_t>( data ); };
+	idEventArg( float data )					{ type = D_EVENT_FLOAT; value = static_cast<intptr_t>( *reinterpret_cast<int *>( &data ) ); };
 	//HUMANHEAD: aob - added const to idVec3
 	idEventArg( const idVec3 &data )			{ type = D_EVENT_VECTOR; value = reinterpret_cast<intptr_t>( &data ); };
 	idEventArg( const idStr &data )				{ type = D_EVENT_STRING; value = reinterpret_cast<intptr_t>( data.c_str() ); };
